requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
pymysql
cryptography>=1.0
PyMySQL>=1.0
